+++ src/Tools/Helpers.ts
@@ -6,26 +6,10 @@ import * as vscode from 'vscode';
  * @param directoryUri URI of base directory
  * @param fileType File type which should be listed
  */
-export async function listFilesOfType(directoryUri: vscode.Uri, fileType: vscode.FileType) {
+async function listFilesOfType(directoryUri: vscode.Uri, fileType: vscode.FileType) {
     return (await vscode.workspace.fs.readDirectory(directoryUri))
         .filter(f => f[1] === fileType)
         .map(d => d[0]);
-}
-
-export async function listDirectories(directoryUri: vscode.Uri) {
-    return listFilesOfType(directoryUri, vscode.FileType.Directory);
-}
-
-export function appendUri(baseUri: vscode.Uri, append: string, autoInsertSeparator: boolean = false, separatorString: string = '/') {
-    const base = baseUri.toString();
-    let usedSeparator = '';
-    if (autoInsertSeparator) {
-        const hasSeparator = base.endsWith(separatorString) || append.startsWith(separatorString);
-        if (!hasSeparator) {
-            usedSeparator = separatorString;
-        }
-    }
-    return vscode.Uri.parse(base + usedSeparator + append);
 }
 
 export function getRelativeUri(baseUri: vscode.Uri, fullUri: vscode.Uri): string | undefined {
@@ -42,11 +26,29 @@ export function getRelativeUri(baseUri: vscode.Uri, fullUri: vscode.Uri): string
     //TODO test
 }
 
+export async function uriExists(uri: vscode.Uri)
+{
+    //TODO test
+    const a = await vscode.workspace.fs.stat(uri);
+    const b = a.type !== vscode.FileType.Unknown;
+}
+
 export function logTimedHeader(message: string, lineLength: number = 150): void {
     const time = new Date().toLocaleTimeString();
     const separator = ' - ';
     const messageWithTime = time + separator + message + ' ';
     const fillLength = lineLength - messageWithTime.length;
     const fill = (fillLength > 0) ? '*'.repeat(fillLength) : '';
-    console.log(messageWithTime + fill);
+    console.warn(messageWithTime + fill);
+}
+
+/**
+ * Pushes an item to an array, only if the item is not null or undefined
+ * @param array Array to which item is pushed
+ * @param item Item which is checked and pushed
+ */
+export function pushDefined<T>(array: T[], item: T | undefined | null) {
+    if (item ?? false) { //TODO test with booleans
+        array.push(item!);
+    }
 }